Butt Organ
Unfinished.

This gives all humanoid species a "butt" organ which is required to fart.
Farting has a chance to cause the ass to fall off.

Includes keybind options for Scream, Fart, Flip, Clap and Spin that each have penalties from overuse.

Includes a Superfart emote which has a low chance to either explode or create an ass rod.